get_eipa_data/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
requests
pandas
pdfplumber
mariadb
mysql-connector-python
openpyxl
numpy
xlsxwriter